Added Spring Security Compatibility
Changes:

Modified the  OneToMany Cascade Types of the Models from MERGE to ALL

Deleted CORS config in favor of a Spring Security approach. Changes can be found in SecurityConfigurer

Removed unused imports, and redundant lines of code in UserController

Modified the endpoints (HTTP verbs and paramets) in UserController to better follow REST

Added the AuthorityDeserizalizer to  properly deserialize user GrantedAuthority when Jackson can maps to Users

Removed createNewUser() from UserController. registerNewUser() in AccountController is meant to be used instead.